remove dedicated middleware dir
