--- backend/api/middleware/cors_middleware.py
@@ -1,18 +1,0 @@
-from config.config import settings
-from fastapi import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
-
-
-def add_cors_middleware(app: FastAPI) -> None:
-    """
-    Adds the cors middleware to the FastAPI app
-
-    :param app: FastAPI app to add the middleware to
-    """
-    app.add_middleware(
-        CORSMiddleware,
-        allow_origins=settings.cors.allow_origins,
-        allow_credentials=settings.cors.allow_credentials,
-        allow_methods=settings.cors.allow_methods,
-        allow_headers=settings.cors.allow_headers,
-    )
