+++ backend/api/v1/aro/endpoints/user.py
@@ -1,10 +1,9 @@
 from uuid import UUID
 
-from data.data_wrappers.wrappers import AROUsersWrapper
-from fastapi import APIRouter
-
 from api.v1.aro.models.auth_requests import UserRequest
 from api.v1.aro.models.auth_responses import AllUsersResponse, UserResponse
+from data.data_wrappers.wrappers import AROUsersWrapper
+from fastapi import APIRouter
 
 aro_user_router = APIRouter(tags=["ARO", "User Information"])
