+++ backend/api/v1/mcc/endpoints/telemetry.py
@@ -1,7 +1,6 @@
+from api.v1.mcc.models.responses import TelemetryListResponse
 from data.data_wrappers.wrappers import TelemetryWrapper
 from fastapi import APIRouter
-
-from api.v1.mcc.models.responses import TelemetryListResponse
 
 telemetry_router = APIRouter(tags=["MCC", "Telemetry"])
